requirements.txt
pytest 
boto3
moto 
pg8000
pandas
pymongo 
pydantic 
bandit 
pip-audit
flake8
black
pytest-cov
python-dotenv